+++ src/main/java/net/geforcemods/securitycraft/items/AdminToolItem.java
@@ -7,6 +7,7 @@ import net.geforcemods.securitycraft.SCContent;
 import net.geforcemods.securitycraft.api.IModuleInventory;
 import net.geforcemods.securitycraft.api.IOwnable;
 import net.geforcemods.securitycraft.api.IPasswordProtected;
+import net.geforcemods.securitycraft.blockentities.DisplayCaseBlockEntity;
 import net.geforcemods.securitycraft.blockentities.SecretSignBlockEntity;
 import net.geforcemods.securitycraft.misc.ModuleType;
 import net.geforcemods.securitycraft.util.PlayerUtils;
@@ -52,6 +53,9 @@ public class AdminToolItem extends Item {
 			BlockEntity be = level.getBlockEntity(pos);
 
 			if (be != null) {
+				if (be instanceof DisplayCaseBlockEntity displayCase && (displayCase.isOpen() && displayCase.getDisplayedStack().isEmpty()))
+					return InteractionResult.PASS;
+
 				boolean hasInfo = false;
 
 				if (be instanceof IOwnable ownable) {