+++ src/main/java/net/geforcemods/securitycraft/items/UniversalBlockModifierItem.java
@@ -3,6 +3,7 @@ package net.geforcemods.securitycraft.items;
 import net.geforcemods.securitycraft.SCContent;
 import net.geforcemods.securitycraft.api.IModuleInventory;
 import net.geforcemods.securitycraft.api.IOwnable;
+import net.geforcemods.securitycraft.blockentities.DisplayCaseBlockEntity;
 import net.geforcemods.securitycraft.blocks.DisguisableBlock;
 import net.geforcemods.securitycraft.inventory.CustomizeBlockMenu;
 import net.geforcemods.securitycraft.util.PlayerUtils;
@@ -36,7 +37,9 @@ public class UniversalBlockModifierItem extends Item {
 		BlockEntity be = level.getBlockEntity(pos);
 		Player player = ctx.getPlayer();
 
-		if (be instanceof IModuleInventory) {
+		if (be instanceof DisplayCaseBlockEntity displayCase && (displayCase.isOpen() && displayCase.getDisplayedStack().isEmpty()))
+			return InteractionResult.PASS;
+		else if (be instanceof IModuleInventory) {
 			if (be instanceof IOwnable ownable && !ownable.isOwnedBy(player)) {
 				if (!(be.getBlockState().getBlock() instanceof DisguisableBlock db) || (((BlockItem) db.getDisguisedStack(level, pos).getItem()).getBlock() instanceof DisguisableBlock))
 					PlayerUtils.sendMessageToPlayer(player, Utils.localize(SCContent.UNIVERSAL_BLOCK_MODIFIER.get().getDescriptionId()), Utils.localize("messages.securitycraft:notOwned", PlayerUtils.getOwnerComponent(ownable.getOwner())), ChatFormatting.RED);